chore: router errors
